// resources/js/types/trans_types.ts
// Trnslation Types 000000000000000000
export type NavBarTranslation = {
    navbar: {
        home: string;
        product: string;
        blog: string;
        about: string;
        contact_us: string;
        account: string;
        website:string;
    };
};
export type HomeTranslation = {
    home: {
        cte_title: string;
        cte_paragraph: string;
        shop_now: string;
        subscribe: string;
        awards: string;
        happy_customers: string;
        sold_products: string;
        cancel: string;
        subscribe_now: string;
        beautiful_scenery:string;
        latest_products:string;
    };
};
export type HomeSec4Types={
    home:{
    sec4_title:string;
    sec4_title1:string;
    sec4_title2:string;
    sec4_title3:string;
    sec4_para1:string;
    sec4_para2:string;
    sec4_para3:string;
    }

}